refactor: separating button component, and deleting redundant functions
