+++ src/components/Button/Edit/index.jsx
@@ -1,25 +1,7 @@
-/* eslint-disable react/prop-types */
-import { useState } from "react";
-import { Modal as ModalFlow } from "flowbite-react";
-import { Link } from "react-router-dom";
-import { formatObs, handleClickReload } from "../../utils/functions";
-import Modal from "../../utils/Modal";
-import { InfoIcon } from "../Icons";
+import Modal from "../../../utils/Modal";
 
-export function RealoadFilters() {
-  return (<button onClick={handleClickReload} className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-md gap-x-1.5 text-sm font-semibold  ring-1 ring-inset ring-gray-300">
-    Limpar Filtros
-  </button>);
-}
-export function ButtonAgenda({ row }) {
-  return (<Link to={`/agenda/${row}`} target="_blank">
-    <button className="bg-cyan-500 hover:bg-cyan-700 text-white py-2 px-4 rounded-md text-sm font-semibold ring-1 ring-inset ring-cyan-300">
-      Agenda
-    </button>
-  </Link>);
-}
 
-export function ButtonEdit({ row }) {
+export default function ButtonEdit({ row }) {
   const modal = new Modal();
   /**
    * Abre o Modal de alterção das licenças.
@@ -31,7 +13,7 @@ export function ButtonEdit({ row }) {
     const emitterDateFormatted = convertDateFormat(row.original.data_emissao);
     const dueDateFormatted = convertDateFormat(row.original.data_vencimento);
     const protocolDateFormatted = convertDateFormat(row.original.data_protocolo_orgao);
-    const observationsFormatted = formatObs(row.original.observacoes);
+    const observationsFormatted = row.original.observacoes.trim();
 
     modal.inputs = {
       myIdInput: row.original.id,
@@ -81,41 +63,4 @@ export function ButtonEdit({ row }) {
       Editar
     </button>
   );
-}
-
-
-
-export default function ModalFlowButton({ onDeleteTask }) {
-  const [openModal, setOpenModal] = useState(false);
-  const deleteTask = () => () => {
-    setOpenModal(false);
-    onDeleteTask();
-  };
-  return (<>
-    <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-md text-sm font-semibold ring-1 ring-inset ring-red-300" onClick={() => setOpenModal(true)}>
-      Apagar
-    </button>
-
-    <ModalFlow show={openModal} size="sm" onClose={() => setOpenModal(false)} popup>
-      <ModalFlow.Header />
-      <ModalFlow.Body>
-        <div className="text-center">
-          <div className="flex justify-center items-center pb-5">
-            <InfoIcon />
-          </div>
-          <h3 className="mb-5 text-lg font-normal text-gray-500 dark:text-gray-400">
-            Deseja deletar essa rotina?
-          </h3>
-          <div className="flex justify-center gap-4">
-            <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-md text-sm font-semibold ring-1 ring-inset ring-red-300" onClick={deleteTask()}>
-              {"Sim, eu quero"}
-            </button>
-            <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-md text-sm font-semibold ring-1 ring-inset ring-blue-300" onClick={() => setOpenModal(false)}>
-              Não, cancele
-            </button>
-          </div>
-        </div>
-      </ModalFlow.Body>
-    </ModalFlow>
-  </>);
 }
